Reverted moving SqsMessage to new package
